fix: GPSArtDetail 화면에서 사용하지 않는 코드 Legacy로 이동 및 GPSArtHomeView ViewModel 적용
