+++ Outline/Outline/Source/Manager/RunningDataManager.swift
@@ -39,7 +39,7 @@ class RunningDataManager: ObservableObject {
     @Published var score: Int = 0
     @Published var num: Int = 0
     
-    @Published var isSaveNewScore = false
+//    @Published var isSaveNewScore = false
     
     @MainActor @Published private(set) var activityID: String?
     @MainActor @Published private(set) var activityToken: String?
@@ -261,7 +261,7 @@ class RunningDataManager: ObservableObject {
                    CourseScoreModel().createOrUpdateScore(courseId: course.id, score: self.score) { scoreResult in
                        switch scoreResult {
                        case .success:
-                           self.isSaveNewScore = true
+//                           self.isSaveNewScore = true
                            print("Score updated successfully")
                        case .failure(let error):
                            print("Error updating score: \(error)")
