+++ Outline/Outline/Source/View/GPSArtHome/DetailView/CardDetailInformationMapView.swift
