+++ Outline/Outline/Source/View/GPSArtHome/DetailView/CardDetailView.swift
